requirements.txt
alembic==1.14.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.0
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
